Implementación completa de animaciones, organización de código para su reutilización y agregar elementos necesarios en la vista de juego
