+++ package.json
@@ -16,11 +16,13 @@
     "@react-three/fiber": "^8.16.2",
     "@react-three/rapier": "^1.1.2",
     "@types/three": "^0.163.0",
-    "ecctrl": "^1.0.77",
+    "ecctrl": "^1.0.75",
+    "firebase": "^10.12.0",
     "leva": "0.9.34",
     "r3f-perf": "6.5.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-router-dom": "^6.23.1",
     "react-scripts": "^5.0.1",
     "react-useanimations": "^2.0.8",
     "three": "^0.163.0",
