+++ public/assets/plato-gato/plato-gato.glb
