+++ public/index.html
@@ -6,7 +6,7 @@
     <title>vr-cat</title>
 </head>
 <body>
-   <div id="root"></div>         
-   <script type="module" src="../src/index.jsx"></script>
+   <div id="root">       
+   <script type="module" src="../src/index.js"></script>
 </body>
 </html>
