+++ src/firebase/firebase.config.js
@@ -1,30 +1,31 @@
 // Import the functions you need from the SDKs you need
 import { initializeApp } from "firebase/app";
+import { getAuth } from "firebase/auth";
+import { getFirestore } from "firebase/firestore";
+// import { getAnalytics } from "firebase/analytics";
+// import { getStorage } from "firebase/storage";
+// import { getDatabase } from "firebase/database";
+
+
 // TODO: Add SDKs for Firebase products that you want to use
 // https://firebase.google.com/docs/web/setup#available-libraries
 
-import {getAuth} from "firebase/auth";
-import {getFirestore} from "firebase/firestore";
-
-
 // Your web app's Firebase configuration
 const firebaseConfig = {
-  apiKey: "AIzaSyB05ilY6TtnJctpY_JBqQvX9Hz6__RgHU8",
-  authDomain: "vr-cat.firebaseapp.com",
-  projectId: "vr-cat",
-  storageBucket: "vr-cat.appspot.com",
-  messagingSenderId: "398421404500",
-  appId: "1:398421404500:web:71f1e893d7a1b28247b6c1"
+  apiKey: "AIzaSyCdLIM9hiWWP_D_PAlCAiwI35JcJXqKv3s",
+  authDomain: "vr-cat-a8b3a.firebaseapp.com",
+  projectId: "vr-cat-a8b3a",
+  storageBucket: "vr-cat-a8b3a.appspot.com",
+  messagingSenderId: "16540522175",
+  appId: "1:16540522175:web:5615aa1564ca725a73180a"
 };
 
 // Initialize Firebase
 const app = initializeApp(firebaseConfig);
 
 
-//export auth
-const auth = getAuth(app);
-
-//export firestore
+const auth =getAuth(app);
 const db = getFirestore(app);
 
-export {auth, db};
+export { auth, db};
+
