+++ src/index.js
@@ -1,0 +1,7 @@
+import React from 'react'
+import ReactDOM from 'react-dom/client'
+import Experience from './Experience'
+import './styles.css'
+
+const root = ReactDOM.createRoot(document.getElementById('root'))
+root.render(<Experience />);
